requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
babel==2.17.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.4.1
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django_components==0.139.1
djangorestframework==3.15.2
djc_core_html_parser==1.0.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
Markdown==3.7
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
zope.interface==7.2
